+++ 1.19.2/Common/src/main/java/com/hypherionmc/craterlib/core/platform/CompatUtils.java
@@ -9,5 +9,8 @@ public interface CompatUtils {
 
     boolean isPlayerActive(BridgedPlayer player);
     String getSkinUUID(BridgedPlayer player);
+    boolean isPlayerBleeding(BridgedPlayer player);
+    boolean playerBledOut(BridgedPlayer player);
+    boolean playerRevived(BridgedPlayer player);
 
 }