+++ 1.20.4/NeoForge/src/main/java/com/hypherionmc/craterlib/CraterLib.java
@@ -2,6 +2,7 @@ package com.hypherionmc.craterlib;
 
 import com.hypherionmc.craterlib.api.events.client.LateInitEvent;
 import com.hypherionmc.craterlib.common.NeoForgeServerEvents;
+import com.hypherionmc.craterlib.compat.PlayerReviveEvents;
 import com.hypherionmc.craterlib.compat.Vanish;
 import com.hypherionmc.craterlib.core.event.CraterEventBus;
 import com.hypherionmc.craterlib.core.networking.CraterPacketNetwork;
@@ -31,6 +32,10 @@ public class CraterLib {
 
         if (ModloaderEnvironment.INSTANCE.isModLoaded("vmod")) {
             NeoForge.EVENT_BUS.register(new Vanish());
+        }
+
+        if (ModloaderEnvironment.INSTANCE.isModLoaded("playerrevive")) {
+            NeoForge.EVENT_BUS.register(new PlayerReviveEvents());
         }
     }