+++ 1.19.3/changelog.md
@@ -7,3 +7,6 @@
 
 - Swapped Built In Config screen for a Cloth Config System, so client side mods can have in-game configs
 - Added new APIs for Simple RPC (V4 rewrite)
+- Added new APIs for getting player information from Commands
+- Added Compat API for Player Revive Mod
+- Added new APIs to know when the Minecraft Whitelist changes